[moe] refactor routers and fix moe bugs with activation checkpoint
