+++ colossalai/nn/layer/moe/__init__.py
@@ -1,8 +1,9 @@
 from .experts import Experts, FFNExperts, TPExperts
-from .layers import MoeLayer, Top1Router, Top2Router, MoeModule
+from .routers import MoeRouter, Top1Router, Top2Router
+from .layers import MoeLayer, MoeModule
 from .utils import NormalNoiseGenerator, UniformNoiseGenerator, build_ffn_experts
 
 __all__ = [
-    'Experts', 'FFNExperts', 'TPExperts', 'Top1Router', 'Top2Router', 'MoeLayer', 'NormalNoiseGenerator',
+    'Experts', 'FFNExperts', 'TPExperts', 'MoeRouter', 'Top1Router', 'Top2Router', 'MoeLayer', 'NormalNoiseGenerator',
     'UniformNoiseGenerator', 'build_ffn_experts', 'MoeModule'
 ]
