+++ tests/test_moe/test_moe_zero_init.py
@@ -19,21 +19,36 @@ from colossalai.utils import get_current_device
 from tests.test_zero.common import CONFIG
 
 
-class MoeModel(CheckpointModule):
+class MoeModel(nn.Module):
 
     def __init__(self, checkpoint: bool = False):
-        super().__init__(checkpoint)
-        self.proj1 = nn.Linear(4, 16)
-        expert_cls = nn.Linear
-        expert_args_dict = dict(in_features=16, out_features=16)
-        self.moe = MoeModule(dim_model=16, num_experts=8, use_residual=True, expert_cls=expert_cls, **expert_args_dict)
-        self.proj2 = nn.Linear(16, 4)
+
+        class TestSubModule(CheckpointModule):
+
+            def __init__(self):
+                super().__init__(checkpoint)
+                expert_cls = nn.Linear
+                expert_args_dict = dict(in_features=16, out_features=16)
+                self.moe = MoeModule(dim_model=16,
+                                     num_experts=8,
+                                     use_residual=True,
+                                     expert_cls=expert_cls,
+                                     **expert_args_dict)
+                self.proj = nn.Linear(16, 4)
+
+            def _forward(self, x):
+                x, y = self.moe(x)
+                x = self.proj(x)
+                return x, y
+
+        super().__init__()
+        self.test_embed = nn.Linear(4, 16)
+        self.test_transform = TestSubModule()
 
     def forward(self, x):
-        x = self.proj1(x)
-        x = self.moe(x)
-        x = self.proj2(x)
-        return x
+        x = self.test_embed(x)
+        x, y = self.test_transform(x)
+        return dict(input=x, moe_loss=y)
 
 
 @parameterize("init_device_type", ['cpu', 'cuda'])
